+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# [Changelog v0.2.0 &nbsp; <img src="https://img.shields.io/badge//Developing-c7a510">](https://github.com/digas99/duo-explained/releases/tag/v0.2.0)
+
+### New Features
+- Added Extension UI language selection
+
 # [Changelog v0.1.2 &nbsp; <img src="https://img.shields.io/badge/Latest-0a7fa6">](https://github.com/digas99/duo-explained/releases/tag/v0.1.2)
 Released on 20/03/2025